opendialog for confirm payment, sends to stripe dashboard
