+++ server/api/StripeResource.py
@@ -25,16 +25,16 @@ class StripeResource(Resource):
     def post(self):
         try:
             data = json.loads(request.data)
-            
+
             chef_id = data.get("chefId")
             user_id = data.get("userId")
             arrival_ts = data.get("arrivalDateTimeStamp")
             ordered_items = data.get("orderedItems")
-            
+
             ret_dict = total_amount_with_meal_objects(ordered_items)
             total_amount_dollars = ret_dict.get("total_amount")
             meal_objects = ret_dict.get("meal_objects")
-            
+
             order = Order(chef_id, user_id, arrival_ts, meal_objects)   
             order.save()
 
@@ -42,7 +42,7 @@ class StripeResource(Resource):
                 amount=int(total_amount_dollars*100),
                 currency='usd'
             )
-            
+
             return jsonify({
                 'clientSecret': intent['client_secret'],
                 'totalAmount': total_amount_dollars,
@@ -51,10 +51,3 @@ class StripeResource(Resource):
 
         except Exception as e:
             return jsonify(error=str(e)), 403
-
-    def get(self):
-        q_params = request.args
-        date = q_params.get("date")
-        datetime_obj = datetime.fromtimestamp(int(int(date)/1000))
-        print(datetime_obj.strftime("%m/%d/%Y, %H:%M"))
-        return datetime_obj.timestamp()
